+++ src/Shrun/Data/Legend.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE UndecidableInstances #-}
-{-# OPTIONS_GHC -Wno-redundant-constraints #-}
 
 -- | Provides types for the legend.
 --